requirements.txt
GoogleAppEngineCloudStorageClient==1.8.3.1
# Additional requirements go here